Continued dev moving ingestion to snovault.
