+++ snovault/loadxl.py
@@ -20,7 +20,7 @@ from pyramid.response import Response
 from pyramid.view import view_config
 from snovault.util import debug_log
 
-from .project_defs import app_project
+from .project_app import app_project
 from .server_defaults import add_last_modified
 
 
@@ -336,7 +336,6 @@ def load_all_gen(testapp, inserts, docsdir, overwrite=True, itype=None, from_jso
             use_itype = True if (itype and isinstance(itype, str)) else False
         else:  # cannot get the file
             err_msg = 'Failure loading inserts from %s. Could not find matching file or directory.' % inserts
-            # import pdb; pdb.set_trace()
             print(err_msg)
             yield str.encode('ERROR: %s\n' % err_msg)
             return
@@ -367,7 +366,6 @@ def load_all_gen(testapp, inserts, docsdir, overwrite=True, itype=None, from_jso
             err_msg = 'No items found in %s' % inserts
         if itype:
             err_msg += ' for item type(s) %s' % itype
-        # import pdb; pdb.set_trace()
         print(err_msg)
         yield str.encode('ERROR: %s' % err_msg)
         return
@@ -436,7 +434,6 @@ def load_all_gen(testapp, inserts, docsdir, overwrite=True, itype=None, from_jso
                               ''.format(a_type, str(first_fields), str(e)))
                         # remove newlines from error, since they mess with generator output
                         e_str = str(e).replace('\n', '')
-                        # import pdb; pdb.set_trace()
                         yield str.encode('ERROR: %s\n' % e_str)
                         return
                         # raise StopIteration
@@ -473,7 +470,6 @@ def load_all_gen(testapp, inserts, docsdir, overwrite=True, itype=None, from_jso
                       a_type, str(an_item), str(e)))
                 print('Full error: %s' % traceback.format_exc())
                 e_str = str(e).replace('\n', '')
-                # import pdb; pdb.set_trace()
                 yield str.encode('ERROR: %s\n' % e_str)
                 return
                 # raise StopIteration
