+++ snovault/project_app.py
@@ -1,0 +1,3 @@
+from dcicutils.project_utils import C4Project
+
+app_project = C4Project.app_project_maker()
